refactor: sessionとdeclarative_baseの分割
