+++ mpt_app/db/session.py
@@ -1,5 +1,5 @@
 from sqlalchemy.ext.asyncio import create_async_engine, AsyncSession
-from sqlalchemy.orm import sessionmaker, declarative_base
+from sqlalchemy.orm import sessionmaker
 
 ASYNC_DB_URL = "mysql+aiomysql://root@db:3306/mpt?charset=utf8mb4"
 
@@ -7,8 +7,6 @@ async_engine = create_async_engine(ASYNC_DB_URL, echo=True)
 async_session = sessionmaker(
     autocommit=False, autoflush=False, bind=async_engine, class_=AsyncSession
 )
-
-Base = declarative_base()
 
 async def get_db():
     async with async_session() as session:
